+++ rector.php
@@ -6,6 +6,7 @@ use Rector\Config\RectorConfig;
 
 return RectorConfig::configure()
     ->withPaths([
+        __DIR__.'/features/bootstrap',
         __DIR__.'/src',
         __DIR__.'/tests',
     ])